+++ src/main/java/com/jd/repository/VideoGameRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface VideoGameRepository extends JpaRepository<VideoGame, Long> {
+public interface VideoGameRepository extends JpaRepository<VideoGame, Integer> {
     // You can define custom queries if needed, but JpaRepository provides common methods by default
 
     @Query("SELECT v FROM VideoGame v")
